Wasn't compiling
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@6799 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/InternalTools/Phones/PhoneEmpDefaults.cs
@@ -88,7 +88,7 @@ namespace OpenDentBusiness{
 				return;
 			}
 			string command="UPDATE phoneempdefault "
-				+"SET StatusOverride = "+POut.Int(PhoneEmpStatusOverride.None)+" "
+				+"SET StatusOverride = "+POut.Int((int)PhoneEmpStatusOverride.None)+" "
 				+"WHERE PhoneExt="+POut.Int(extension)+" "
 				+"AND EmployeeNum="+POut.Long(empNum);
 			Db.NonQ(command);
